+++ features/admin.connections.v1/components/edit/settings/authenticator-settings.tsx
@@ -1005,6 +1005,7 @@ export const AuthenticatorSettings: FunctionComponent<IdentityProviderSettingsPr
 
             return (
                 <AuthenticatorFormFactory
+                    connectionId={ identityProvider.id }
                     connectionSettingsMetaData={ connectionSettingsMetaData }
                     mode={ AuthenticatorSettingsFormModes.EDIT }
                     authenticator={ authenticator }