+++ features/admin.applications.v1/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @wso2is/admin.applications.v1
 
+## 2.43.17
+
+### Patch Changes
+
+- [#10656](https://github.com/wso2/identity-apps/pull/10656) [`021521ffd4fd9cd38872723c06826739f88ee548`](https://github.com/wso2/identity-apps/commit/021521ffd4fd9cd38872723c06826739f88ee548) Thanks [@savindi7](https://github.com/savindi7)! - Use product name config in Try it app.
+
+- Updated dependencies [[`fd9aa87db360d2d96fdf310b97a65896324323b1`](https://github.com/wso2/identity-apps/commit/fd9aa87db360d2d96fdf310b97a65896324323b1)]:
+  - @wso2is/admin.claims.v1@2.31.5
+
 ## 2.43.16
 
 ### Patch Changes