+++ features/admin.connections.v1/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @wso2is/admin.connections.v1
 
+## 2.40.0
+
+### Minor Changes
+
+- [#10608](https://github.com/wso2/identity-apps/pull/10608) [`c9071eeebca69a453584392d0c6fea505a733b98`](https://github.com/wso2/identity-apps/commit/c9071eeebca69a453584392d0c6fea505a733b98) Thanks [@VIHANGAGIT](https://github.com/VIHANGAGIT)! - Add support for `select` fields in connection create and edit templates, including a `useDynamicFieldOptions` hook that resolves field options from existing connections declared via `optionsSource` metadata.
+
+### Patch Changes
+
+- Updated dependencies [[`c9071eeebca69a453584392d0c6fea505a733b98`](https://github.com/wso2/identity-apps/commit/c9071eeebca69a453584392d0c6fea505a733b98)]:
+  - @wso2is/forms@2.11.0
+  - @wso2is/i18n@2.45.6
+
 ## 2.39.1
 
 ### Patch Changes